refresh default route when app changed

+++ test/unittest/services/nci_ce_proxy_test/nci_ce_proxy_test.cpp
@@ -61,7 +61,7 @@ HWTEST_F(NciCeProxyTest, SetCeHostListener001, TestSize.Level1)
     std::shared_ptr<NCI::INciCeInterface::ICeHostListener> listener = nullptr;
     std::shared_ptr<NciCeProxy> nciCeProxy = std::make_shared<NciCeProxy>();
     nciCeProxy->SetCeHostListener(listener);
-    bool computeRoutingParams = nciCeProxy->ComputeRoutingParams(KITS::DefaultPaymentType::TYPE_ESE);
+    bool computeRoutingParams = nciCeProxy->ComputeRoutingParams(static_cast<int>(KITS::DefaultPaymentType::TYPE_ESE));
     ASSERT_TRUE(computeRoutingParams == true);
 }
 
@@ -73,7 +73,7 @@ HWTEST_F(NciCeProxyTest, SetCeHostListener001, TestSize.Level1)
 HWTEST_F(NciCeProxyTest, ComputeRoutingParams001, TestSize.Level1)
 {
     std::shared_ptr<NciCeProxy> nciCeProxy = std::make_shared<NciCeProxy>();
-    bool computeRoutingParams = nciCeProxy->ComputeRoutingParams(KITS::DefaultPaymentType::TYPE_ESE);
+    bool computeRoutingParams = nciCeProxy->ComputeRoutingParams(static_cast<int>(KITS::DefaultPaymentType::TYPE_ESE));
     ASSERT_TRUE(computeRoutingParams == true);
 }
